Initial pass at removing unnecessary environment annotation bytecode hacks to reduce the number of unnecessary references to fabric
