+++ src/main/java/com/wildfire/main/cloud/SyncingTooFrequentlyException.java
@@ -18,9 +18,6 @@
 
 package com.wildfire.main.cloud;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-
-@Environment(EnvType.CLIENT)
+/// @apiNote Only use this on the client side
 public class SyncingTooFrequentlyException extends RuntimeException {
 }
