+++ src/main/java/com/wildfire/render/WildfireModelRenderer.java
@@ -24,15 +24,13 @@ import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.wildfire.main.uvs.UVDirection;
 import com.wildfire.main.uvs.UVLayout;
 import com.wildfire.main.uvs.UVQuad;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import org.joml.Matrix3f;
 import org.joml.Matrix4f;
 import org.joml.Vector3f;
 import org.joml.Vector3fc;
 import org.joml.Vector4f;
 
-@Environment(EnvType.CLIENT)
+/// @apiNote Only use this on the client side
 public final class WildfireModelRenderer {
     private WildfireModelRenderer() {
         throw new UnsupportedOperationException();
@@ -86,12 +84,12 @@ public final class WildfireModelRenderer {
             float f = x + dx;
             float f1 = y + dy;
             float f2 = z + dz;
-            x = x - delta;
-            y = y - delta;
-            z = z - delta;
-            f = f + delta;
-            f1 = f1 + delta;
-            f2 = f2 + delta;
+            x -= delta;
+            y -= delta;
+            z -= delta;
+            f += delta;
+            f1 += delta;
+            f2 += delta;
 
             initQuads(tW, tH, dx, dy, dz, quads,
                     new PositionTextureVertex(f, y, z, 0.0F, 8.0F),
